backend/pyproject.toml
[project]
name = "prelegal-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiofiles>=24.1",
    "litellm>=1.40",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
